Build Backend and UI

+++ sicotyc-server/routes/userRole-routes.js
@@ -5,10 +5,12 @@
 const { Router } = require('express');
 const { check } = require('express-validator');
 const { validateFields } = require('../middlewares/validate-fields');
-const { createUserRole, deleteAllUserRole, deleteUserRole } = require('../controllers/userRole-controller');
+const { getRolesByUserId, createUserRole, deleteAllUserRole, deleteUserRole } = require('../controllers/userRole-controller');
 const { validateJWT } = require('../middlewares/validate-jwt');
 
 const router = Router();
+
+router.get('/:uid', getRolesByUserId);
 
 router.post('/',
     [
